+++ R/sergeant.r
@@ -300,6 +300,35 @@ drill_version <- function(drill_con) {
   }
 }
 
+#' Show all the available Drill built-in functions & UDFs
+#'
+#' @md
+#' @param drill_con drill server connection object setup by \code{drill_connection()}
+#' @note You _must_ be using Drill 1.15.0+ to use this function
+#' @export
+#' @family Dill direct REST API Interface
+#' @references \href{https://drill.apache.org/docs/}{Drill documentation}
+#' @examples \dontrun{
+#' drill_connection() %>% drill_functions()
+#' }
+drill_functions <- function(drill_con) {
+
+  stopifnot(utils::compareVersion(drill_version(drill_con), "1.15.0") >= 0)
+
+  if (inherits(drill_con, "src_drill")) {
+    dplyr::collect(
+      dplyr::tbl(drill_con, dplyr::sql("(SELECT * FROM sys.functions)"))
+    )
+  } else {
+    drill_query(
+      drill_con = drill_con,
+      query = "SELECT * FROM sys.functions",
+      uplift = TRUE,
+      .progress = FALSE
+      )
+  }
+}
+